beautify and some comments
