+++ task_04/src/dijkstra.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <limits>
+#include <map>
+#include <vector>
+
+std::vector<double> Dijkstra(std::vector<std::map<int, double>> data, int source);
