+++ task_05/README.md
@@ -1,1 +1,1 @@
-# Реализовать решение задач RMQ. Чем более эфективный алгоритм, тем больше балов.
+# Реализовать решение задач RMQ. Чем более эффективный алгоритм, тем больше баллов.
