+++ task_06/README.md
@@ -1,1 +1,1 @@
-# Реализовать решение задач LCA. Чем более эфективный алгоритм, тем больше балов.
+# Реализовать решение задач LCA. Чем более эффективный алгоритм, тем больше баллов.
